+++ FastPbfReader/Structs/GpsNode.cs
@@ -6,11 +6,11 @@ namespace OsmFastPbf
   public struct GpsNode
   {
     public readonly long id;
-    public readonly long latCode;
-    public readonly long lonCode;
+    public readonly int latCode;
+    public readonly int lonCode;
     public double Latitude { get { return latCode / 10000000.0; } }
     public double Longitude { get { return lonCode / 10000000.0; } }
-    public GpsNode(long id, long latCode, long lonCode)
+    public GpsNode(long id, int latCode, int lonCode)
     {
       this.id = id;
       this.latCode = latCode;
